feat: add memory profiling and change lint
