+++ test-requirements.txt
@@ -1,6 +1,8 @@
 mypy>=1,<2
 black>=22.8.0,<24
+openpyxl>=3.1.2<4
 pre-commit>=2.20.0,<4
 pytest>=7.1.3
 pytest-benchmark>=4.0.0,<5
 ruff>=0.0.138,<0.1
+xlrd>=2.0.1<3
